Move to summarizerNode subfolder or internal module
