+++ packages/runtime/runtime-utils/src/index.ts
@@ -10,7 +10,8 @@ export * from "./objectstoragepartition";
 export * from "./objectstorageutils";
 export * from "./requestParser";
 export * from "./serializer";
-export * from "./summarizerNodeContract";
+// eslint-disable-next-line import/no-internal-modules
+export * from "./summarizerNode/index";
 export * from "./summaryTracker";
 export * from "./summaryUtils";
 export * from "./utils";
